LC-palindrome-number [Runtime: 0 ms; Memory: 8.6 MB]

+++ math/palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        int temp = x;
+        long temp = x;
         long num = 0;
         while(temp>0){
             num = num*10 + temp % 10;
